requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.3
djangorestframework==3.15.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.0
rsa==4.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
